adding LCM check for servers other than pullserver

+++ rsBasePrep.ps1
@@ -494,20 +494,21 @@ Function Install-DSC {
       Start-sleep -Seconds 10
    }
    while ( $systemProcess.processName.count -gt 1 )#>
-   do {
-      if(!(Test-Path -Path "C:\Windows\System32\Configuration\Current.mof") -or !(Test-Path -Path "C:\Windows\System32\Configuration\Pending.mof")) {
-         Write-EventLog -LogName DevOps -Source BasePrep -EntryType Information -EventId 1002 -Message "Current.mof has not yet been created and Pending.mof does not exist."
-         if((Get-ScheduledTask -TaskName "Consistency").State -eq "Ready") {
-            Write-EventLog -LogName DevOps -Source BasePrep -EntryType Information -EventId 1002 -Message "Consistency task is not running and no Current.mof file exists, restarting Consistency task."
-            taskkill /F /IM WmiPrvSE.exe
-            Get-ScheduledTask -TaskName "Consistency" | Start-ScheduledTask
+   if($role -ne "Pull") {
+      do {
+         if(!(Test-Path -Path "C:\Windows\System32\Configuration\Current.mof") -or !(Test-Path -Path "C:\Windows\System32\Configuration\Pending.mof")) {
+            Write-EventLog -LogName DevOps -Source BasePrep -EntryType Information -EventId 1002 -Message "Current.mof has not yet been created and Pending.mof does not exist."
+            if((Get-ScheduledTask -TaskName "Consistency").State -eq "Ready") {
+               Write-EventLog -LogName DevOps -Source BasePrep -EntryType Information -EventId 1002 -Message "Consistency task is not running and no Current.mof file exists, restarting Consistency task."
+               taskkill /F /IM WmiPrvSE.exe
+               Get-ScheduledTask -TaskName "Consistency" | Start-ScheduledTask
+            }
+            Write-EventLog -LogName DevOps -Source BasePrep -EntryType Information -EventId 1002 -Message "Starting to sleep and will recheck status of LCM."
+            Start-Sleep -Seconds 30
          }
-         Write-EventLog -LogName DevOps -Source BasePrep -EntryType Information -EventId 1002 -Message "Starting to sleep and will recheck status of LCM."
-         Start-Sleep -Seconds 30
       }
+      while(!(Test-Path -Path "C:\Windows\System32\Configuration\Current.mof"))
    }
-   while(!(Test-Path -Path "C:\Windows\System32\Configuration\Current.mof"))
-   
    
    
    ### Pullserver specific tasks, install WindowsFeature Web-Service, install SSL certificates then run rsEnvironments.ps1 to install DSC
